fix(comms): comments, clean code; add permissions test
