+++ .changes/communication_firewall.md
@@ -1,0 +1,6 @@
+---
+"stronghold-communication": minor
+"iota-stronghold": minor
+---
+
+Implement a configurable firewall in the communication actor, add a macro to derive permissions for requests.
