+++ communication/src/actor/connections.rs
@@ -33,7 +33,7 @@ impl ConnectionManager {
         }
     }
 
-    // Check is peer is in the curently active connections.
+    // Check is peer is in the currently active/ allowed connections.
     pub fn is_active_connection(&self, peer_id: &PeerId) -> bool {
         self.map.get(peer_id).is_some()
     }
